% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extendShinySemantic}
\alias{extendShinySemantic}
\title{Add \code{fomantic.plus} Dependencies to \code{shiny.semantic} Application}
\usage{
extendShinySemantic()
}
\value{
A \code{shiny.tag.list} containing tags to enable the JS and CSS required for this package.
}
\description{
In order for any of the \code{fomantic.plus} functionality to work

This will be automatically included in any \code{xxx_page} function in this package,
for example \code{\link{navbar_page}}.
}
\examples{
if (interactive()) {
  library(shiny)
  library(shiny.semantic)
  library(fomantic.plus)

  ui <- semanticPage(
    title = "Hello Fomantic UI",
    tags$head(
      extendShinySemantic()
    )
  )
}

}
