% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{field_validation}
\alias{field_validation}
\alias{field_rule}
\title{Field Validation for Fomantic UI}
\usage{
field_validation(id, ..., extra_params = NULL)

field_rule(rule, prompt = NULL, value = NULL)
}
\arguments{
\item{id}{HTML id of the field to be validated}

\item{...}{A series of \code{field_rule}s that will be applied to the field}

\item{extra_params}{A named list of extra parameters that can be added to the field validation. For example
\code{optional = TRUE} means the field will only be checked if non-empty}

\item{rule}{The type of rule to be applied. Valid rules are available in Details.}

\item{prompt}{Text to be displayed in the UI if the validation fails. Leave \code{NULL} if want to use default text.}

\item{value}{Certain fields require a value to check validation. Check Details if the \code{rule} requires a value.}
}
\description{
A field validation assigns a series of rules that have been assigned to a particular input and checks, upon
the form submission, whether or not the input meets all specified criteria.
}
\details{

}
\examples{
# E-mail validations
field_validation("email", field_rule("email"))

# Password validation
field_validation(
  "password",
  field_rule("empty"),
  field_rule("minLength", value = 8),
  field_rule("regExp", "Must contain at least one special character", "\\\\W")
)

}
\references{
\url{https://fomantic-ui.com/behaviors/form.html}
}
\seealso{
\code{\link{form_validation}}
}
