% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{f_grid}
\alias{f_grid}
\alias{f_row}
\alias{f_column}
\title{Create Fomantic UI Grid}
\usage{
f_grid(..., class = "")

f_row(..., class = "")

f_column(..., class = "")
}
\arguments{
\item{...}{Other arguments to be added as attributes of the tag (e.g. style, class or childrens etc.)}

\item{class}{Additional classes to add to the grid elements}
}
\description{
Using Fomantic UI's implementation of grid with the use of rows and columns.
}
\details{
For a full list of classes to give to grid elements, visit
\url{https://fomantic-ui.com/collections/grid.html}
}
\examples{
# Simple two column grid
f_grid(
  f_row(
    class = "two column",
    f_column(),
    f_column()
  )
)

}
