% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\docType{data}
\name{fui_el}
\alias{fui_el}
\title{Create Fomantic UI Elements}
\format{
An object of class \code{list} of length 41.
}
\usage{
fui_el
}
\arguments{
\item{...}{Tag attributes (named arguments) and children (unnamed arguments).
A named argument with an \code{NA} value is rendered as a boolean attributes (see example).
Children may include any combination of:
\itemize{
\item{Other tags objects}
\item{\code{\link[htmltools]{HTML}()} strings}
\item{\code{\link[htmltools]{htmlDependency}()}s}
\item{Single-element atomic vectors}
\item{\code{list()}s containing any combination of the above}
}}

\item{emoji}{\code{fui_el$emoji} - The string of the emoji name}

\item{country}{\code{fui_el$flag} - Wither the country name or 2 character ISO code}

\item{icon}{\code{fui_el$icon} - The space separated name of the Font Awesome icon}

\item{html_tag}{For certain elements, such as headers and items, multiple HTML tags can be used.
The default is set to \code{div}, but can be set to any valid HTML tag.}
}
\description{
Create an R object that represents a Fomantic UI Element e.g. segment or container. The contents have remained
as minimal as possible to enable the greatest possibile flexibility.
}
\examples{
# List
fui_el$list(
  fui_el$item("Item 1"),
  fui_el$item("Item 2"),
  fui_el$item("Item 3")
)

# Pink Segment
fui_el$segment(
  class = "pink"
)

# Grid
fui_el$grid(
  fui_el$row(
    class = "two column",
    fui_el$column(),
    fui_el$column()
  )
)

}
\seealso{
\url{https://fomantic-ui.com} for stlying Fomantic UI elements, \code{\link[htmltools]{builder}}
}
\keyword{datasets}
