% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbarPage.R
\name{update_navbar_page}
\alias{update_navbar_page}
\title{Change the selected tab}
\usage{
update_navbar_page(
  session = shiny::getDefaultReactiveDomain(),
  id,
  selected = NULL
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{id}{The id of the navbar object}

\item{target}{The tab value to show}
}
\description{
Change the selected tab on the client
}
\examples{
if (interactive()) {
  library(shiny)
  library(shiny.semantic)
  library(fomantic.plus)

  ui <- navbar_page(
    title = "App Title",
    id = "navbar",
    tab_panel(
      "Plot",
      action_button("select", "Go to Table"),
      value = "plot"
    ),
    tab_panel("Summary", value = "summary"),
    tab_panel("Table", value = "table")
  )

  server <- function(input, output, session) {
    observeEvent(input$select, update_navbar_page(session, "navbar", "table"))
  }

  shinyApp(ui, server)
}

}
